renamed Search to SearchResidents
